requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile requirements.in
#
astroid==4.0.1
    # via pylint
attrs==25.4.0
    # via
    #   outcome
    #   trio
autopep8==2.3.2
    # via -r requirements.in
boto3==1.40.66
    # via -r requirements.in
botocore==1.40.66
    # via
    #   boto3
    #   s3transfer
botocore-stubs==1.40.66
    # via types-boto3
certifi==2025.10.5
    # via
    #   requests
    #   selenium
charset-normalizer==3.4.4
    # via requests
dill==0.4.0
    # via pylint
feedgen==1.0.0
    # via -r requirements.in
h11==0.16.0
    # via wsproto
idna==3.11
    # via
    #   requests
    #   trio
iniconfig==2.3.0
    # via pytest
isort==7.0.0
    # via pylint
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
lxml==6.0.2
    # via feedgen
mccabe==0.7.0
    # via pylint
mypy==1.18.2
    # via -r requirements.in
mypy-extensions==1.1.0
    # via mypy
outcome==1.3.0.post0
    # via
    #   trio
    #   trio-websocket
packaging==25.0
    # via pytest
pathspec==0.12.1
    # via mypy
platformdirs==4.5.0
    # via pylint
pluggy==1.6.0
    # via pytest
pycodestyle==2.14.0
    # via autopep8
pygments==2.19.2
    # via pytest
pylint==4.0.2
    # via -r requirements.in
pysocks==1.7.1
    # via urllib3
pytest==9.0.2
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   botocore
    #   feedgen
pytz==2025.2
    # via user-agent
requests==2.32.5
    # via -r requirements.in
s3transfer==0.14.0
    # via boto3
selenium==4.38.0
    # via -r requirements.in
six==1.17.0
    # via
    #   python-dateutil
    #   user-agent
sniffio==1.3.1
    # via trio
sortedcontainers==2.4.0
    # via trio
sqlalchemy==2.0.44
    # via -r requirements.in
tomlkit==0.13.3
    # via pylint
trio==0.32.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.12.2
    # via selenium
types-awscrt==0.28.2
    # via botocore-stubs
types-boto3==1.40.66
    # via -r requirements.in
types-python-dateutil==2.9.0.20251008
    # via -r requirements.in
types-requests==2.32.4.20250913
    # via -r requirements.in
types-s3transfer==0.14.0
    # via types-boto3
types-yt-dlp==2025.12.8.20251210
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   mypy
    #   selenium
    #   sqlalchemy
urllib3[socks]==2.5.0
    # via
    #   botocore
    #   requests
    #   selenium
    #   types-requests
user-agent==0.1.14
    # via -r requirements.in
websocket-client==1.9.0
    # via selenium
websockets==15.0.1
    # via types-yt-dlp
wsproto==1.2.0
    # via trio-websocket
yt-dlp==2025.10.22
    # via -r requirements.in
